+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 [comment]: # ( Copyright Contributors to the Open Cluster Management project )
 # Release Content
 
--  Add in `cm attach cluster` and `cm create cluster` supports for custom labels
+- Add in `cm attach cluster` and `cm create cluster` supports for custom labels
+- Make the output yamls more pretty
+- import the `init`, `join`, `accept` and `get token` commands from [clusteradm](https://github.com/open-cluster-management-io/clusteradm)
+- Add `cm enable addons`